Feature/fixing clang gnu compiler warnings (#45)
* Support for CMake builds. Added Linux stub port for now to allow cross
compilation.

* Fixing misuse of doxygen documentation

* Converting prints from %lu to %u where possible as per request in comments related to (#35)

* Fixing clang compiler warnings.

* Fixing build test to work regardless of configuration.

* Adding in for Zynq port dependency on uncached_memory target.

* Excluding build_test from all build - must request it explicitly.

* Ensuring XPAR_XSDPS_0_IS_CACHE_COHERENT is 1 for this build since using
a modified xsdps driver that requires it.

* Updating compiler warnings.

* Fixing documentation, signed/unsigned conversion. Bugfix for time based code for removing failure in 2032 due to uint32_t used for time.

* Fixing error creation using FF_createERR everywhere missed.

* Fixing build_test in ci.yml

* Uncrustify: triggered by comment.

---------

+++ portable/common/ff_ramdisk.c
@@ -419,10 +419,10 @@ BaseType_t FF_RAMDiskShowPartition( FF_Disk_t * pxDisk )
         FF_PRINTF( "Partition Nr   %8u\n", pxDisk->xStatus.bPartitionNumber );
         FF_PRINTF( "Type           %8u (%s)\n", pxIOManager->xPartition.ucType, pcTypeName );
         FF_PRINTF( "VolLabel       '%8s' \n", pxIOManager->xPartition.pcVolumeLabel );
-        FF_PRINTF( "TotalSectors   %8lu\n", ( unsigned long ) pxIOManager->xPartition.ulTotalSectors );
-        FF_PRINTF( "SecsPerCluster %8lu\n", ( unsigned long ) pxIOManager->xPartition.ulSectorsPerCluster );
-        FF_PRINTF( "Size           %8lu KB\n", ( unsigned long ) ulTotalSizeKB );
-        FF_PRINTF( "FreeSize       %8lu KB ( %d perc free )\n", ( unsigned long ) ulFreeSizeKB, iPercentageFree );
+        FF_PRINTF( "TotalSectors   %8u\n", ( unsigned ) pxIOManager->xPartition.ulTotalSectors );
+        FF_PRINTF( "SecsPerCluster %8u\n", ( unsigned ) pxIOManager->xPartition.ulSectorsPerCluster );
+        FF_PRINTF( "Size           %8u KB\n", ( unsigned ) ulTotalSizeKB );
+        FF_PRINTF( "FreeSize       %8u KB ( %d perc free )\n", ( unsigned ) ulFreeSizeKB, iPercentageFree );
     }
 
     return xReturn;
